Clean up Kevin's fractal detail implementation: - Replaced the manual mip level calculation with Texture2D.CalculateLevelOfDetail. This uses the hardware to calculate mip levels and is quad-uniform. Thus it is no longer necessary to calculate derivatives and use SampleGradient, which frees up several full-precision registers. - Switched blending between the three levels in the fractal details to use a gaussian blending function that preserves variance. This is technically incorrect as the contents of the detail maps aren't gaussian, but they are close enough that it works better than the old average function that causes the details to look faded. - Switched non-fractal detail keyword name to _FRACTAL_DETAILS_OFF instead of _DETAILS_UV_ON, and made _DETAILS_ON a separate shader feature. Globally enabling _FRACTAL_DETAILS_OFF will force off fractal details on all materials! - Fixed LitMAS GUI handling of the detail keywords. The original dropdown did not use a BaseMaterialField class so it could not set the paired _detail float property, and did not dirty the material after setting the keywords. The MaterialIntPopup class is now extended to allow enabling and disabling lists of keywords with each option, and the detail mode dropdown now uses this. Additionally, the "off" option was removed and the detail header's on/off checkbox was restored. - Moved the volumetric and volumetric hq keywords into the default lit variants include. I am unsure why these weren't there in the first place? Also disabled the hq keyword on mobile for now as quest does not ever use it.

+++ Shaders/LitMAS/LitMASInclude/ShaderInjector/Injections/Injection_Emission.hlsl
@@ -14,8 +14,8 @@ TEXTURE2D(_EmissionMap);
 	{
 		emission += SAMPLE_TEXTURE2D(_EmissionMap, sampler_BaseMap, uv_main) * _EmissionColor;
 		emission.rgb *= lerp(albedo.rgb, half3(1, 1, 1), emission.a);
-		half emNoV = _EmissionFalloff >= 0 ? abs(fragData.NoV) : 1.0 - abs(fragData.NoV);
+		half emNoV = _EmissionFalloff >= half(0) ? abs(fragData.NoV) : half(1.0) - abs(fragData.NoV);
 		emission.rgb *= saturate(pow(emNoV, abs(_EmissionFalloff)));
-		emission = max(emission,0);
+		emission = max(emission,half(0));
 	}
 //#!INJECT_END
